+++ App/common/common.h
@@ -72,7 +72,7 @@ void writeBackupRegister(ENUM_backupRegister backupRegisterId, uint32_t value);
 uint32_t readBackupRegister(ENUM_backupRegister backupRegisterId);
 
 uint32_t SYS_GetAdc(int channel);
-uint16_t SYS_GetVoltage(int channel, uint32_t adcValue);
+int32_t SYS_GetVoltage(int channel, uint32_t adcValue);
 
 int getDecimal(float value, int digits);
